refactor mtp.py
